Detect Linux attrs applicable to restore target
This is a somewhat quick-and-dirty take on finding the set of Linux
attrs that the restore-target file system accepts.  Whereas previously
applying attrs to files was an "all or nothing" operation (if any attr
was not applicable on the file system, none would be applied), this
patch enables us to apply the supported attrs only.

The following things need work:
- whereas previously we'd emit error messages, this patch causes us to
  silently skip attrs.  There's a TODO without a good plan in _helpers.c
- should we somehow restore across multiple file systems (which I'm not
  sure bup could handle at all anyway?), we would currently not detect
  their differences: we only detect applicable attrs for the first path
  we restore

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>