Read basic pack files without git cat-file
This patch adds git.CatFile as a drop-in "replacement" for git.CatPipe,
and makes vfs.py use it in place of CatPipe.  As it relies on
git.PackIdxList for object lookups, it enables us to experiment with
midx-backed pack reading.

The patch also changes the signature of git._decode_packobj(), which
should be ok as it didn't turn out to have any callers yet.  I
experimented a bit with the original _decode_packobj, which would be
passed a buffer() slice of an mmapped pack-file, and found high memory
consumption and seemingly very "greedy" disk reads slowed things down
terribly.  Hence this change, using file.read() on packs.

Note this is very much work in progress; see the various TODOs in
CatFile().

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>