git._decode_packobj: don't overread so much
This is a small silly change, but it improved my restore speed by about
5%, so it's worth committing.  We previously used chunkyreader() to read
from pack files, but that tends to grab chunks that, at 64k, are way
bigger than a typical hashsplitted and compressed chunk.  We now just do
8k reads, which turns out to make a real difference (cProfile attributes
all of the saved time specifically to file.read()).

Note that 8k is just a lucky pick - one might want to try other values.

There's also an adjustment to tgit.py, because the new optimisation
relies on the fact that pack files end with a checksum footer, so that
decompress() always has some left-overs.  The test-"pack" originally
lacked a footer.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>