Revert earlier "disable midx/bloom" commit.
My introducing a skip_bloom switch was based on a misunderstanding of
what the already present do_bloom does, so let's get rid of it.  Also,
we didn't need the skip_midx changes anymore since we now do by default
choose to use the midxes again, therefore, revert the lot!

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>